Snap: Map61B.java, MyHashMap.java, TestMyHashMap.java

+++ lab8/hashmap/MyHashMap.java
@@ -167,7 +167,7 @@ public class MyHashMap<K, V> implements Map61B<K, V> {
      */
     @Override
     public void put(K key, V value) {
-        double division = size / initialSize;
+        double division = ((double)size / initialSize);
         if (division > loadFactor) {
             resize(initialSize * 2);
         }
